--- frontend/src/app/components/overview/cloud-overview/cloud-overview.html
@@ -1,42 +1,0 @@
-<div class="column pad borders greedy">
-    <h2 class="row heading3 content-end">
-        <span class="greedy">Cloud Resources</span>
-        <a class="link" data-bind="
-            text: cloudResourceCount,
-            href: { route: 'pools', params: { tab: 'cloud' } }
-        "></a>
-    </h2>
-    <hr>
-    <div class="row push-prev">
-        <div class="column content-middle greedy">
-            <svg-icon class="icon-medium push-next"
-                params="name: awsResourceIcon"
-                data-bind="tooltip: 'AWS S3 resources'"
-            ></svg-icon>
-            <span class="remark highlight" data-bind="
-                text: awsResourceCount
-            "></span>
-        </div>
-
-        <div class="column content-middle greedy">
-            <svg-icon class="icon-medium push-next"
-                params="name: azureResourceIcon"
-                data-bind="tooltip: 'Azure resources'"
-            ></svg-icon>
-            <span class="remark highlight" data-bind="
-                text: azureResourceCount
-            "></span>
-        </div>
-
-        <div class="column content-middle greedy">
-            <svg-icon  class="icon-medium push-next"
-                params="name: genericResourceIcon"
-                data-bind="tooltip: 'S3 compatible resources'"
-            ></svg-icon>
-            <span class="remark highlight" data-bind="
-                text: genericResourceCount
-            "></span>
-        </div>
-    </div>
-</div>
-
