--- frontend/src/app/components/overview/cloud-overview/cloud-overview.less
@@ -1,2 +1,0 @@
-cloud-overview {
-}
