--- frontend/src/app/components/overview/cluster-overview/cluster-overview.html
@@ -1,25 +1,0 @@
-<div class="column pad borders greedy">
-    <h2 class="heading3 row content-middle">
-        <span class="greedy">Cluster</span>
-        <a class="link preview-only" data-bind="
-            text: serverCountText,
-            href: { route: 'cluster' }
-        "></a>
-    </h2>
-    <hr>
-    <div class="row greedy content-middle">
-        <svg-icon class="push-next"
-            params="name: statusIcon().name"
-            data-bind="css: statusIcon().css"
-        ></svg-icon>
-        <span class="highlight" data-bind="text: statusText"></span>
-    </div>
-
-    <div class="row greedy content-middle">
-        <svg-icon class="push-next"
-            params="name: highAvailabilityIcon().name"
-            data-bind="css: highAvailabilityIcon().css"
-        ></svg-icon>
-        <span class="highlight" data-bind="text: highAvailabilityText"></span>
-    </div>
-</div>
