--- frontend/src/app/components/overview/cluster-overview/cluster-overview.less
@@ -1,2 +1,0 @@
-cluster-overview {
-}
