+++ frontend/src/app/components/overview/overview-panel/overview-panel.html
@@ -1,9 +1,5 @@
 <div class="row content-box push-both">
-    <pools-overview class="column c4 push-next"></pools-overview>
-    <section class="column c4 push-both">
-        <cloud-overview class="column c4"></cloud-overview>
-        <cluster-overview class="column c4 push-both"></cluster-overview>
-        <system-health class="preview-only column c4"></system-health>
-    </section>
+    <resource-overview class="column c4 push-next"></resource-overview>
+    <system-health class="column c4"></system-health>
     <buckets-overview class="column c4 push-prev"></buckets-overview>
 </div>
