+++ frontend/src/app/components/overview/resource-overview/resource-overview.less
@@ -1,4 +1,4 @@
-pools-overview {
+resource-overview {
     display: block;
 
     /*
@@ -8,6 +8,17 @@ pools-overview {
     background: url('@{assets}/stars.jpg');
     background-size: cover;
 
+    .resource-icon {
+        margin-bottom: @gutter/2;
+    }
+
+    .nodes-pool-icon {
+        fill: @color8;
+    }
+
+    .nodes-legend .legend-caption {
+        font-size: @font-size3;
+    }
 
     .nodes-chart {
         margin: auto;
