+++ frontend/src/app/components/overview/system-health/system-health.less
@@ -1,3 +1,7 @@
 system-health {
     display: block;
+
+    h3 {
+        font-size: @font-size3;
+    }
 }
